+++ packages/liveblocks-core/src/__tests__/room.test.ts
@@ -30,7 +30,7 @@ import { ClientMsgCode } from "../protocol/ClientMsg";
 import type { BaseMetadata } from "../protocol/Comments";
 import { OpCode } from "../protocol/Op";
 import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
-import { CrdtType } from "../protocol/SerializedCrdt";
+import { CrdtType, nodeStreamToCompactNodes } from "../protocol/SerializedCrdt";
 import { ServerMsgCode } from "../protocol/ServerMsg";
 import type { RoomConfig, RoomDelegates } from "../room";
 import { createRoom } from "../room";
@@ -983,8 +983,9 @@ describe("room", () => {
     wss.onConnection((conn) => {
       conn.server.send(
         serverMessage({
-          type: ServerMsgCode.STORAGE_STATE,
-          items: [["root", { type: CrdtType.OBJECT, data: {} }]],
+          type: ServerMsgCode.STORAGE_CHUNK,
+          done: true,
+          nodes: [["root", CrdtType.OBJECT, {}]],
         })
       );
     });
@@ -1017,12 +1018,12 @@ describe("room", () => {
       wss.onConnection((conn) => {
         conn.server.send(
           serverMessage({
-            type: ServerMsgCode.STORAGE_STATE,
-            items: [["root", { type: CrdtType.OBJECT, data: {} }]],
-            //                                              ^^
-            //                                   NOTE: Storage is initially
-            //                                   empty, so initialStorage keys
-            //                                   will get added
+            type: ServerMsgCode.STORAGE_CHUNK,
+            done: true,
+            nodes: [["root", CrdtType.OBJECT, {}]],
+            //                                ^^
+            //                        NOTE: Storage is initially empty,
+            //                        so initialStorage keys will get added
           })
         );
       });
@@ -1052,8 +1053,9 @@ describe("room", () => {
     wss.onConnection((conn) => {
       conn.server.send(
         serverMessage({
-          type: ServerMsgCode.STORAGE_STATE,
-          items: [["root", { type: CrdtType.OBJECT, data: { x: 0 } }]],
+          type: ServerMsgCode.STORAGE_CHUNK,
+          done: true,
+          nodes: [["root", CrdtType.OBJECT, { x: 0 }]],
         })
       );
     });
@@ -1155,8 +1157,9 @@ describe("room", () => {
     wss.onConnection((conn) => {
       conn.server.send(
         serverMessage({
-          type: ServerMsgCode.STORAGE_STATE,
-          items: [["root", { type: CrdtType.OBJECT, data: { x: 0 } }]],
+          type: ServerMsgCode.STORAGE_CHUNK,
+          done: true,
+          nodes: [["root", CrdtType.OBJECT, { x: 0 }]],
         })
       );
     });
@@ -1263,8 +1266,9 @@ describe("room", () => {
     wss.onConnection((conn) => {
       conn.server.send(
         serverMessage({
-          type: ServerMsgCode.STORAGE_STATE,
-          items: [["root", { type: CrdtType.OBJECT, data: { x: 0 } }]],
+          type: ServerMsgCode.STORAGE_CHUNK,
+          done: true,
+          nodes: [["root", CrdtType.OBJECT, { x: 0 }]],
         })
       );
     });
@@ -1301,8 +1305,9 @@ describe("room", () => {
     wss.onConnection((conn) => {
       conn.server.send(
         serverMessage({
-          type: ServerMsgCode.STORAGE_STATE,
-          items: [["root", { type: CrdtType.OBJECT, data: { x: 0 } }]],
+          type: ServerMsgCode.STORAGE_CHUNK,
+          done: true,
+          nodes: [["root", CrdtType.OBJECT, { x: 0 }]],
         })
       );
     });
@@ -1385,8 +1390,9 @@ describe("room", () => {
       wss.onConnection((conn) => {
         conn.server.send(
           serverMessage({
-            type: ServerMsgCode.STORAGE_STATE,
-            items: [["root", { type: CrdtType.OBJECT, data: { x: 0 } }]],
+            type: ServerMsgCode.STORAGE_CHUNK,
+            done: true,
+            nodes: [["root", CrdtType.OBJECT, { x: 0 }]],
           })
         );
       });
@@ -1908,8 +1914,9 @@ describe("room", () => {
       wss.onConnection((conn) =>
         conn.server.send(
           serverMessage({
-            type: ServerMsgCode.STORAGE_STATE,
-            items: newInitStorage,
+            type: ServerMsgCode.STORAGE_CHUNK,
+            done: true,
+            nodes: Array.from(nodeStreamToCompactNodes(newInitStorage)),
           })
         )
       );