+++ packages/liveblocks-core/src/crdts/LiveObject.ts
@@ -14,6 +14,7 @@ import type {
 import { OpCode } from "../protocol/Op";
 import type {
   IdTuple,
+  NodeStream,
   SerializedChild,
   SerializedCrdt,
   SerializedObject,
@@ -77,7 +78,7 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
   public static detectLargeObjects = false;
 
   static #buildRootAndParentToChildren(
-    items: IdTuple<SerializedCrdt>[]
+    items: NodeStream
   ): [root: SerializedRootObject, nodeMap: ParentToChildNodeMap] {
     const parentToChildren: ParentToChildNodeMap = new Map();
     let root: SerializedRootObject | null = null;
@@ -105,11 +106,11 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
 
   /** @private Do not use this API directly */
   static _fromItems<O extends LsonObject>(
-    items: IdTuple<SerializedCrdt>[],
+    nodes: NodeStream,
     pool: ManagedPool
   ): LiveObject<O> {
     const [root, parentToChildren] =
-      LiveObject.#buildRootAndParentToChildren(items);
+      LiveObject.#buildRootAndParentToChildren(nodes);
     return LiveObject._deserialize(
       ["root", root],
       parentToChildren,