+++ packages/liveblocks-core/src/index.ts
@@ -289,6 +289,13 @@ export type {
   UserRoomSubscriptionSettings,
 } from "./protocol/RoomSubscriptionSettings";
 export type {
+  CompactChildNode,
+  CompactListNode,
+  CompactMapNode,
+  CompactNode,
+  CompactObjectNode,
+  CompactRegisterNode,
+  CompactRootNode,
   IdTuple,
   SerializedChild,
   SerializedCrdt,
@@ -298,14 +305,14 @@ export type {
   SerializedRegister,
   SerializedRootObject,
 } from "./protocol/SerializedCrdt";
-export { CrdtType } from "./protocol/SerializedCrdt";
+export { CrdtType, nodeStreamToCompactNodes } from "./protocol/SerializedCrdt";
 export type {
   BroadcastedEventServerMsg,
   CommentsEventServerMsg,
   RejectedStorageOpServerMsg,
   RoomStateServerMsg,
   ServerMsg,
-  StorageStateServerMsg,
+  StorageChunkServerMsg,
   UpdatePresenceServerMsg,
   UpdateStorageServerMsg,
   UserJoinServerMsg,