+++ packages/liveblocks-redux/src/__tests__/index.test.ts
@@ -7,7 +7,13 @@ import type {
   ServerMsg,
   UpdatePresenceServerMsg,
 } from "@liveblocks/core";
-import { ClientMsgCode, OpCode, ServerMsgCode } from "@liveblocks/core";
+import {
+  ClientMsgCode,
+  CrdtType,
+  nodeStreamToCompactNodes,
+  OpCode,
+  ServerMsgCode,
+} from "@liveblocks/core";
 import type { Reducer } from "@reduxjs/toolkit";
 import { configureStore } from "@reduxjs/toolkit";
 import { http, HttpResponse } from "msw";
@@ -181,8 +187,9 @@ async function prepareWithStorage<T extends Record<string, unknown>>(
 
   socket.callbacks.message[0]!({
     data: JSON.stringify({
-      type: ServerMsgCode.STORAGE_STATE,
-      items: options.items,
+      type: ServerMsgCode.STORAGE_CHUNK,
+      done: true,
+      nodes: Array.from(nodeStreamToCompactNodes(options.items)),
     }),
   } as MessageEvent);
 
@@ -239,8 +246,9 @@ describe("middleware", () => {
 
     socket.callbacks.message[0]!({
       data: JSON.stringify({
-        type: ServerMsgCode.STORAGE_STATE,
-        items: [obj("root", {})],
+        type: ServerMsgCode.STORAGE_CHUNK,
+        done: true,
+        nodes: [["root", CrdtType.OBJECT, {}]],
       }),
     } as MessageEvent);
 
@@ -279,9 +287,7 @@ describe("middleware", () => {
           targetActor: -1,
           data: { cursor: { x: 0, y: 0 } },
         },
-        {
-          type: ClientMsgCode.FETCH_STORAGE,
-        },
+        { type: ClientMsgCode.FETCH_STORAGE },
       ]);
 
       await waitFor(() => socket.sentMessages[1] != null);
@@ -309,9 +315,7 @@ describe("middleware", () => {
           targetActor: -1,
           data: { cursor: { x: 0, y: 0 } },
         },
-        {
-          type: ClientMsgCode.FETCH_STORAGE,
-        },
+        { type: ClientMsgCode.FETCH_STORAGE },
       ]);
     });
 
@@ -330,9 +334,7 @@ describe("middleware", () => {
           targetActor: -1,
           data: { cursor: { x: 0, y: 0 } },
         },
-        {
-          type: ClientMsgCode.FETCH_STORAGE,
-        },
+        { type: ClientMsgCode.FETCH_STORAGE },
       ]);
 
       store.dispatch({ type: "SET_CURSOR", cursor: { x: 1, y: 1 } });
@@ -362,9 +364,7 @@ describe("middleware", () => {
           targetActor: -1,
           data: { cursor: { x: 0, y: 0 } },
         },
-        {
-          type: ClientMsgCode.FETCH_STORAGE,
-        },
+        { type: ClientMsgCode.FETCH_STORAGE },
       ]);
 
       store.dispatch(leaveRoom());